+++ NEWS.md
@@ -1,3 +1,15 @@
+# rcompendium 0.5.1
+
+After first submission to CRAN:
+
+* `DESCRIPTION` file: replace "The aim of the package 'rcompendium' is to make 
+easier..." by "Makes easier..."
+* Function `set_credentials()` does not write the `.Rprofile` file anymore 
+(not allowed by CRAN policies). Instead this function opens this file and users 
+need to manually paste the content of the clipboard.
+* Replace `utils::installed.packages()` by `find.package()`
+
+
 # rcompendium 0.5
 
 * Check for valid package name